backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.9
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
requests==2.28.2
supabase==1.0.3
stripe==5.4.0
openai==0.27.4
aiohttp==3.8.4
aiodns==3.0.0
pytest==7.3.1
httpx==0.24.0
tenacity==8.2.2
sentry-sdk==1.19.1
alembic==1.10.3
psycopg2-binary==2.9.6
redis==4.5.4
celery==5.2.7
flower==1.2.0
pillow==9.5.0